Use CMD in Dockerfile
According to Dockerfile best practices
(https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#entrypoint)
the best use for ENTRYPOINT is to set the image’s main command and then
use CMD as the default flags. This makes it easier to change the flags
passed to monerod when running the Docker image.